this repository contains some prototype versions and design documentation for a syscall intermediation project.
the bulk of this commit is a highly-bodged version of the moss open source linux-workalike kernel for aarch64.
since our plan is just to use this as a translation layer between linux syscalls and a distributed systems layer, we
kinda went to town ripping things apart - in particular the crate distinction between kernel and libkernel.

the current status is that we were able to start bringing up a less-munged version of moss and print messages using
an hvc, but not start a child process or handle any syscalls